// libs/administrative-boundary/ui/src/lib/adapters/department-vm-props.interface.ts
import { Status } from '@cmz/administrative-boundary-domain';
import { ActionDropdownItem } from '@cmz/shared-ui';
import { StatusStyle } from '../enums/status-style.enum';

export interface DepartmentVmProps {
    uniqId: string;
    code: string;
    name: string;
    description: string;
    regionId: string;
    regionName: string;
    populationSize: number;
    infrastructureCount: number;
    municipalitiesCount: number;
    status: Status;
    statusLabel: string;
    statusStyle: StatusStyle;
    updatedAt: string;
    actionsRef: string;
    dropdownActions: ActionDropdownItem[];
    disableDropdown: boolean;
    tooltipDropdown: string;
}